pyproject.toml
[project]
name = "exoplanet-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.7",
    "openai>=2.31.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.0",
]
